config/env.py
from dotenv import load_dotenv
import os

load_dotenv()

def get_env(key: str, default=None):
    value = os.getenv(key, default)
    if value is None:
        raise RuntimeError(f"Missing required env var: {key}")
    return value
