requirements.txt
langchain
langchain-community
langchain-groq
langgraph
faiss-cpu
sentence-transformers
ragas
trulens-eval
zenml
python-dotenv